feat(logger): remove wrap + almost finished with new features
