--- packages/paperdave-logger/example.mjs
@@ -1,7 +1,0 @@
-import log, { createLogger } from './dist/index.js';
-
-log.error(
-  'Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World Hello World '
-);
-log.error(new Error('error'));
-log.error(new TypeError('other error'));
