+++ packages/paperdave-logger/package.json
@@ -20,8 +20,7 @@
   "dependencies": {
     "ansi-escapes": "^5.0.0",
     "chalk": "4",
-    "strip-ansi": "^7.0.1",
-    "wrap-ansi": "^8.0.1"
+    "strip-ansi": "^7.0.1"
   },
   "devDependencies": {
     "@paperdave/utils": "workspace:^1.1.2",
