+++ packages/paperdave-logger/src/filter.ts
@@ -1,0 +1,7 @@
+export function setLogFilter(...filter: Array<string | string[]>) {
+  // TODO: implement
+}
+
+export function isLogVisible(id: string) {
+  return true;
+}
