+++ packages/paperdave-logger/src/index.ts
@@ -1,10 +1,11 @@
 import * as log from './log';
 
+export * from './deprecated';
 export { CLIError, type PrintableError } from './error';
 export { injectLogger } from './inject';
 export { LogLevel, setLevel } from './level';
 export * from './log';
-export * from './log-base';
+export { createLogger } from './log-base';
 export {
   BarStyle,
   Progress,
@@ -14,6 +15,7 @@ export {
   type WithProgressOptions,
 } from './progress';
 export { Spinner, withSpinner, type SpinnerOptions } from './spinner';
+export type { CustomLoggerColor, CustomLoggerOptions, LogFunction, StringLike } from './types';
 export { isUnicodeSupported, logSymbols } from './unicode';
 export { LogWidget } from './widget';
 export default log;
